requirements.txt
aiogram==2.25
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
pyaes==1.6.1
pyasn1==0.5.1
PySocks==1.7.1
python-socks==2.4.4
pytz==2023.3.post1
rsa==4.9
six==1.16.0
Telethon==1.33.1
tzlocal==5.2
yarl==1.9.4
